requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
Django==5.0.1
django-restframework==0.0.1
django-websockets==0.0.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gevent==23.9.1
greenlet==3.0.3
grpcio==1.60.0
inflection==0.5.1
install==1.3.5
msgpack==1.0.7
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
setuptools==69.0.3
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
websocket==0.2.1
websockets==11.0.1
zope.event==5.0
zope.interface==6.1
